+++ atn-atm-mediator/src/common/mod.rs
@@ -2,3 +2,4 @@ pub mod config;
 pub mod did_conversion;
 pub mod errors;
 pub mod jwt_auth;
+pub mod statistics;